added suggestion from internal demo
